Invoke-CMApplyDriverPackage.ps1 version 2.2.5
New deployment type named 'Prestaged' for use with Prestaged Media (OEMMedia) where driver packages may or may not be included on the media. This method will verify the package version (default behaviour for Prestaged Media deployments) for prestaged/cached packages and download newer version if applicable, or will download package if it is not prestaged/cached. Drivers will be applied in same way as BareMetal deployment type.